LayoutTests/imported/w3c/web-platform-tests/navigation-api/navigation-methods/return-value/traverseTo-detach-same-document-before-navigate-event-expected.txt



FAIL traverseTo() promise rejections when detaching an iframe before onnavigate (same-document) assert_equals: expected 1 but got 0
